rename radio_client hil to rfcore
